+++ wizer/configuration.py
@@ -1,3 +1,7 @@
+from wizer.best_sections.fastest import get_fastest_section
+from wizer.best_sections.climb import get_best_climb_section
+
+
 # supported for parsing
 supported_formats = {"gpx", "fit"}
 
@@ -44,14 +48,34 @@ avg_attributes = {
 # configuration of best sections
 rank_limit = 3
 
-# fastest section to parse in activities (in kilometer, only integers allowed)
-fastest_sections = {
-    1,
-    2,
-    3,
-    5,
-    10,
-}
+# best sections config
+best_sections = [
+    {
+        "kind": "fastest",
+        "parser": get_fastest_section,
+        "distances": [  # in meter
+            1_000,
+            2_000,
+            3_000,
+            5_000,
+            10_000,
+        ],
+    },
+    {
+        "kind": "climb",
+        "parser": get_best_climb_section,
+        "distances": [  # in meter
+            100,
+            200,
+            500,
+            1_000,
+            2_000,
+        ],
+    },
+]
+available_best_section = [sec["kind"] for sec in best_sections]
+fastest_distances = best_sections[0]["distances"]
+climb_distances = best_sections[1]["distances"]
 
 # with respect to the table listing activities used both on dashboard and sport page
 number_of_rows_per_page_in_table = 40