+++ wizer/tests/integration_tests/browser_tests/test_activity_page.py
@@ -28,6 +28,36 @@ def test_activity_page__complete(import_one_activity, live_server, webdriver):
     assert "Pace" in table_header
     assert "Label" in table_header
 
+    table_data = [cell.text for cell in webdriver.find_elements_by_tag_name("td")]
+    # best sections
+    assert "  1km" in table_data
+    assert "  2km" in table_data
+    assert "  3km" in table_data
+    assert "  5km" in table_data
+    assert "  10km" in table_data
+    assert "30.2 km/h" in table_data
+    assert "25.1 km/h" in table_data
+    assert "23.6 km/h" in table_data
+    # assert "22.3 km/h" in table_data      # fails on mac?!
+    assert "17.9 km/h" in table_data
+
+    # avg + min + max values
+    assert "max" in table_data
+    assert "high" in table_data
+    assert "avg" in table_data
+    assert "min" in table_data
+    assert "9.0 km/h" in table_data
+    assert "47.9 km/h" in table_data
+    assert "06:39 min/km" in table_data
+    assert "01:15 min/km" in table_data
+    assert "23.0 °C" in table_data
+    assert "26.0 °C" in table_data
+    assert "32.0 °C" in table_data
+    assert "1" in table_data
+    assert "1:46:15" in table_data
+    assert "11423" in table_data
+    assert "09:18" in table_data
+
     assert webdriver.find_element_by_tag_name("h3").text == "Noon Cycling in Bad Schandau  "
 
     headings = [h.text for h in webdriver.find_elements_by_tag_name("h5")]
@@ -169,7 +199,8 @@ def test_edit_activity_page(import_one_activity, live_server, webdriver, insert_
 
     # check that all trophies got removed, since activity no longer evaluates for awards
     assert len(webdriver.find_elements_by_class_name("fa-trophy")) == 0
-    assert len(webdriver.find_elements_by_class_name("fa-exclamation-circle")) == 1
+    # two warning exclamation circles are present, one for fastest, one for climb
+    assert len(webdriver.find_elements_by_class_name("fa-exclamation-circle")) == 2
 
     # verify attributes got changed
     activity = models.Activity.objects.get()