+++ wizer/tests/integration_tests/test_views.py
@@ -1,7 +1,10 @@
 from django.urls import reverse
 import pytest
+import numpy as np
 
 from wizer import models
+from wizer import configuration
+from wizer.views import get_flat_list_of_pks_of_activities_in_top_awards
 
 
 def test_help_view(db, client):
@@ -25,7 +28,7 @@ def test_settings_view(db, client):
 def test_best_sections_view(db, client):
     response = client.get(reverse("awards"))
     assert response.status_code == 200
-    assert "Your Awards" in response.content.decode("UTF-8")
+    assert "awards" in response.content.decode("UTF-8")
 
 
 def test_activity_view__activity_present(db, client, settings, sport, activity):
@@ -37,3 +40,9 @@ def test_activity_view__activity_present(db, client, settings, sport, activity):
 def test_activity_view__no_activity(db, client):
     with pytest.raises(models.Activity.DoesNotExist):
         client.get("/activity/1")
+
+
+def test_get_flat_list_of_pks_of_activities_in_top_awards(db, import_demo_data):
+    result_pks = get_flat_list_of_pks_of_activities_in_top_awards(configuration.rank_limit)
+    assert len(result_pks) == 7
+    assert len(np.unique(result_pks)) == 7